requirements.txt
pandas~=1.4.1
psycopg2-binary
ipython-sql